refactor(main.rs): move work into lib and add tests
